feat: Implemented "move a folder" from current authenticated user.

+++ MiniDropBox.Application/Implementations/FolderService.cs
@@ -38,7 +38,7 @@ namespace MiniDropBox.Application.Implementations
                     CreatedAt = DateTime.UtcNow,
                     Path = isSubFolder
                         ? Path.Combine($"{parentFolder!.Path}", $"{folderDTO.Name}")
-                        : Path.Combine($"{folderDTO.Name}",$"{folderDTO.UserId}")
+                        : Path.Combine($"{folderDTO.Name}", $"{folderDTO.UserId}")
                 };
 
                 if (isSubFolder)
@@ -64,8 +64,29 @@ namespace MiniDropBox.Application.Implementations
 
         public async Task<bool> DeleteFolderAsync(int folderId)
         {
-            var deletedFolder = await _folderRepository.DeleteAsync(folderId);
-            return deletedFolder != null;
+            await _unitOfWork.BeginTransactionAsync();
+
+            try
+            {
+                var folder = await _folderRepository.GetByIdAsync(folderId);
+                if (folder == null)
+                {
+                    return false;
+                }
+
+                var hasChildren = folder.SubFolders != null && folder.SubFolders.Count != 0;
+                if (hasChildren)
+                    return false;
+
+                var deletedFolder = await _folderRepository.DeleteAsync(folderId);
+                await _unitOfWork.CommitAsync();
+                return deletedFolder != null;
+            }
+            catch (Exception)
+            {
+                await _unitOfWork.RollbackAsync();
+                throw;
+            }
         }
 
         public async Task<Folder?> GetFolderByIdAsync(int folderId)
@@ -84,10 +105,115 @@ namespace MiniDropBox.Application.Implementations
             return _treeBuilder.BuildTree(folders);
         }
 
-        public async Task<bool> UpdateFolderAsync(Folder folder)
+        public async Task<Result<string>> MoveFolderAsync(MoveFolderDTO moveFolderDTO, int userId)
         {
-            var updatedFolder = await _folderRepository.UpdateAsync(folder);
-            return updatedFolder != null;
+            await _unitOfWork.BeginTransactionAsync();
+
+            try
+            {
+                var folder = await _folderRepository.GetByIdWithSubfoldersRecursivelyAsync(moveFolderDTO.Id);
+                if (folder == null || folder.UserId != userId)
+                    return Result<string>.Failure("Folder not found or you don't have permission");
+
+                if (folder.Id == moveFolderDTO.NewParentFolderId)
+                    return Result<string>.Failure("Cannot ove a folder into itslef");
+
+                var newParent = await _folderRepository.GetByIdAsync(moveFolderDTO.NewParentFolderId);
+                if (newParent == null || newParent.UserId != userId)
+                    return Result<string>.Failure("New parent folder not found");
+
+                // Avoid circular reference
+                var temp = newParent;
+                while (temp != null)
+                {
+                    if (temp.Id == folder.Id)
+                        return Result<string>.Failure("Cannot move folder into its descendant (circular reference)");
+
+                    temp = temp.ParentFolderId != null
+                        ? await _folderRepository.GetByIdAsync(temp.ParentFolderId.Value)
+                        : null;
+                }
+
+                // Update the folder's parent
+                folder.ParentFolderId = newParent.Id;
+
+                // Update the folder's and subfolders path 
+                folder.Path = Path.Combine(newParent.Path, folder.Name);
+                UpdateSubfolderPathsRecursively(folder);
+
+                await _folderRepository.UpdateAsync(folder);
+                await _unitOfWork.CommitAsync();
+
+                return Result<string>.Success($"Folder moved to: {folder.Path}");
+            }
+            catch (Exception)
+            {
+                await _unitOfWork.RollbackAsync();
+                throw;
+            }
+        }
+
+        public async Task<Result<string>> UpdateFolderAsync(UpdateFolderDTO updateFolderDTO)
+        {
+            await _unitOfWork.BeginTransactionAsync();
+
+            try
+            {
+                var existing = await _folderRepository.GetByIdAsync(updateFolderDTO.Id);
+                if (existing == null)
+                    return Result<string>.Failure("Folder not found");
+
+                if (updateFolderDTO.ParentFolderId == updateFolderDTO.Id)
+                    return Result<string>.Failure("Folder cannot be its own parent");
+
+                Folder? parentFolder = null;
+
+                if (updateFolderDTO.ParentFolderId != null)
+                {
+                    parentFolder = await _folderRepository.GetByIdAsync(updateFolderDTO.ParentFolderId.Value);
+                    if (parentFolder == null)
+                        return Result<string>.Failure("Parent folder not found");
+
+                    // Avoid circular reference
+                    var temp = parentFolder;
+                    while (temp != null)
+                    {
+                        if (temp.Id == existing.Id)
+                            return Result<string>.Failure("Cannot assign a child folder as parent (circular reference)");
+
+                        temp = temp.ParentFolderId !=null
+                            ? await _folderRepository.GetByIdAsync(temp.ParentFolderId.Value)
+                            : null;
+                    }
+                }
+
+                existing.Name = updateFolderDTO.Name;
+                existing.ParentFolderId = updateFolderDTO.ParentFolderId;
+
+                existing.Path = parentFolder != null
+                    ? Path.Combine(parentFolder.Path, updateFolderDTO.Name)
+                    : Path.Combine(updateFolderDTO.Name, existing.UserId.ToString());
+
+                var updatedFolder = await _folderRepository.UpdateAsync(existing);
+                await _unitOfWork.CommitAsync();
+
+                return Result<string>.Success($"Uploaded folder, new path: {updatedFolder!.Path} .");
+            }
+            catch (Exception)
+            {
+                await _unitOfWork.RollbackAsync();
+                throw;
+            }
+        }
+
+        // Auxiliar method to update folder path recursively
+        private void UpdateSubfolderPathsRecursively(Folder folder)
+        {
+            foreach (var sub in folder.SubFolders)
+            {
+                sub.Path = Path.Combine(folder.Path, sub.Name);
+                UpdateSubfolderPathsRecursively(sub);
+            }
         }
     }
 }
